Added challenge achievements and removed the now-redundant achievements_EZ2_client.cpp
