+++ sp/src/common/hl2orange.spa.h
@@ -352,6 +352,9 @@ extern "C" {
 #define ACHIEVEMENT_EZ2_FLIP_APC					217
 #define ACHIEVEMENT_EZ2_SURRENDERED_HEAL			218
 #define ACHIEVEMENT_EZ2_WILSON_CLOSETS				219
+#define ACHIEVEMENT_EZ2_CHALLENGES_ALL_BRONZE		220
+#define ACHIEVEMENT_EZ2_CHALLENGES_ALL_SILVER		221
+#define ACHIEVEMENT_EZ2_CHALLENGES_ALL_GOLD			222
 
 #ifndef _XBOX
 #define ACHIEVEMENT_PORTAL_TRANSMISSION_RECEIVED	149
